Accept black background PR, separate components by function, and clear out promiscuous functions
