+++ app/page.tsx
@@ -1,7 +1,7 @@
 import { getFlags } from "@/app/lib/getFlags";
-import { fetchFlags, fetchFilteredFlags } from "@/app/lib/data";
-import FlagsPage from "@/components/my/flags-page";
-import { FooterFlags } from "@/app/ui/footer/footer-flags";
+import { fetchFlags } from "@/app/lib/data";
+import SortableGallery from "@/app/ui/gallery/sortable-gallery";
+import Gnb from "./ui/gnb/gnb";
 
 export const fetchCache = 'force-no-store';
 
@@ -17,19 +17,18 @@ export default async function Home(props: {
   let flags;
 
   try {
-    flags = await fetchFilteredFlags(query);
+    flags = await fetchFlags(query);
   } catch (error) {
     console.error('fetchFlags failed, using getFlags as fallback:', error);
     flags = await getFlags(); // Fallback to getFlags if fetchFlags fails
   }
 
-  // const flags = await fetchFlags();
-
   return (
-    <>
-      <FlagsPage initialFlags={flags} />
-
-      <FooterFlags copyrightHref="/123" />
-    </>
+    <div className="grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]">
+      <Gnb />
+      <main className="flex flex-col gap-8 row-start-2 items-center sm:items-start">
+        <SortableGallery filteredFlags={flags} />
+      </main>
+    </div>
   );
 }
