--- app/ui/gnb/avatar-sadness.tsx
@@ -1,11 +1,0 @@
-import { Flag } from "lucide-react"
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"
-
-export function AvatarSadness() {
-  return (
-    <Avatar className="hidden md:block">
-      <AvatarImage src="/sadness_2_40x40.webp" />
-      <AvatarFallback><Flag /></AvatarFallback>
-    </Avatar>
-  )
-}
